bugfix - cast int to int64 to be able to do a bitwise and. Solves issue where no exclude_flags are specified for get_fits

+++ mkidcore/pixelflags.py
@@ -56,7 +56,7 @@ class FlagSet(object):
     def bitmask(self, flags, unknown='error'):
         """Return the bitmask corresponding to the flags"""
         if not flags:
-            return 0
+            return np.uint64(0)
         if isinstance(flags, str):
             flags = [flags]
